Change difference function to use Long instead of Integer
Forsee issue with delta of lots of tiles (from images)

+++ src/jwblangley/difference/DifferenceFunction.java
@@ -3,13 +3,13 @@ package jwblangley.difference;
 @FunctionalInterface
 public interface DifferenceFunction<T> {
 
-  Integer apply(T a, T b);
+  Long apply(T a, T b);
 
-  default Integer squareDifference(T a, T b) {
+  default Long squareDifference(T a, T b) {
     return apply(a, b) * apply(a, b);
   }
 
-  default Integer absoluteDifference(T a, T b) {
+  default Long absoluteDifference(T a, T b) {
     return apply(a, b) < 0 ? apply(a, b) * -1 : apply(a, b);
   }
 }
